+++ app-frontend/src/app/components/pages/project/analyses/visualize/index.js
@@ -1,5 +1,5 @@
 import { Map } from 'immutable';
-import { get } from 'lodash';
+import { get, min, max } from 'lodash';
 
 import tpl from './index.html';
 import {createRenderDefinition} from '_redux/histogram-utils';
@@ -7,23 +7,44 @@ import {nodesFromAst, astFromNodes} from '_redux/node-utils';
 
 class AnalysesVisualizeController {
     constructor(
-        $rootScope, $scope, $state, $log,
-        mapService, projectService, analysisService
+        $rootScope, $scope, $state, $log, $q,
+        uuid4, mapService, projectService, analysisService
     ) {
         'ngInject';
         $rootScope.autoInject(this, arguments);
     }
 
     $onInit() {
+        this.isLoadingAnalysis = false;
+        this.isLoadingAnalysisError = false;
         this.selected = new Map();
+        this.analysesMap = new Map();
         this.projectId = this.$state.params.projectId;
         this.layerColorHex = {};
         this.analyses = this.getAnalysisIds().map(analysisId => {
+            // create a trackId for each analysis to enable displaying
+            // same analysis multiple times especially for map layers
+            const trackId = this.uuid4.generate();
+            const analysisPromise = this.mapLayerFromAnalysis(analysisId, trackId);
             return {
                 id: analysisId,
-                analysisTile: this.mapLayerFromAnalysis(analysisId)
+                trackId,
+                analysisTile: analysisPromise.then(a => a.analysisTile),
+                histogram: analysisPromise.then(a => (
+                    {histogram: a.histogram, statistics: a.statistics})),
+                statistics: analysisPromise.then(a => a.statistics)
             };
         });
+
+        this.histogramBounds = this.$q.all(this.analyses.map(a => a.statistics))
+            .then(statistics => {
+                return statistics.reduce((acc, stat) => {
+                    return {
+                        min: min([acc.min, stat.zmin]),
+                        max: max([acc.max, stat.zmax])
+                    };
+                }, {});
+            });
     }
 
     getAnalysisIds() {
@@ -34,34 +55,48 @@ class AnalysesVisualizeController {
         return analysis;
     }
 
-    mapLayerFromAnalysis(analysisId) {
-        return this.analysisService.getAnalysis(analysisId).then(analysis => {
-            this.analysisService.getNodeHistogram(analysisId).then(histogram => {
-                const {
-                    renderDefinition,
-                    histogramOptions
-                } = createRenderDefinition(histogram);
-                const newNodeDefinition = Object.assign(
-                    {}, analysis.executionParameters,
-                    {
-                        metadata: Object.assign({}, analysis.executionParameters.metadata, {
-                            renderDefinition,
-                            histogramOptions
-                        })
-                    }
-                );
-                const nodes = nodesFromAst(analysis.executionParameters);
-                const updatedAnalysis = astFromNodes({analysis, nodes}, [newNodeDefinition]);
-                return this.analysisService.updateAnalysis(updatedAnalysis);
-            });
+    mapLayerFromAnalysis(analysisId, trackId) {
+        this.isLoadingAnalysis = true;
+        this.isLoadingAnalysisError = false;
+        return this.$q.all({
+            analysis: this.analysisService.getAnalysis(analysisId),
+            histogram: this.analysisService.getNodeHistogram(analysisId),
+            statistics: this.analysisService.getNodeStatistics(analysisId)
+        }).then(({analysis, histogram, statistics}) => {
+            this.analysesMap = this.analysesMap.set(trackId, analysis);
             this.setLayerColorHex(analysis);
-            return analysis;
-        }).then((analysis) => {
+            const {
+                renderDefinition,
+                histogramOptions
+            } = createRenderDefinition(histogram);
+            const newNodeDefinition = Object.assign(
+                {}, analysis.executionParameters,
+                {
+                    metadata: Object.assign({}, analysis.executionParameters.metadata, {
+                        renderDefinition,
+                        histogramOptions
+                    })
+                }
+            );
+            const nodes = nodesFromAst(analysis.executionParameters);
+            const updatedAnalysis = astFromNodes({analysis, nodes}, [newNodeDefinition]);
+            this.analysisService.updateAnalysis(updatedAnalysis);
+            return {analysis, histogram, statistics};
+        }).then(({analysis, histogram, statistics}) => {
             const tileUrl = this.analysisService.getAnalysisTileUrl(analysisId);
             return {
-                analysis,
-                mapTile: L.tileLayer(tileUrl, {maxZoom: 30})
+                histogram,
+                statistics,
+                analysisTile: {
+                    analysis,
+                    mapTile: L.tileLayer(tileUrl, {maxZoom: 30})
+                }
             };
+        }).catch(err => {
+            this.$log.error(err);
+            this.isLoadingAnalysisError = true;
+        }).finally(() => {
+            this.isLoadingAnalysis = false;
         });
     }